requirements.txt
flask
gunicorn
mysql.connector
appwrite
logtail-python